+++ src/main/java/roomescape/config/AdminOnlyInterceptor.java
@@ -1,4 +1,4 @@
-package roomescape;
+package roomescape.config;
 
 import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletRequest;
