+++ src/main/java/roomescape/config/WebMvcConfiguration.java
@@ -1,4 +1,4 @@
-package roomescape;
+package roomescape.config;
 
 import java.util.List;
 import org.springframework.context.annotation.Configuration;
